refactor(phase4d): extract shared stream logic to ai/adapters/shared/streamShared.ts
三家 *-aisdk 的 stream.ts 逐字重复部分抽到单一共享件（ThinkTagParser、
基础类型 BaseStreamResult/BaseStreamParams、getProviderConfigFromStore、
accumulateToolCall、emitStreamComplete/emitStreamError/detectAndEmitToolUseTags、
PROMPT_MODE_STOP_SEQUENCES）。各家 fullStream 循环（推理/思考时机差异）保留不动。

去重以满足 SonarCloud 新代码重复率门禁（≤3%），为后续目录收敛（4b/4c）扫清障碍。
零行为改动，type-check / test(32/32) / build 全绿。
