+++ src/shared/api/anthropic-aisdk/stream.ts
@@ -6,135 +6,32 @@
 import { streamText, generateText } from 'ai';
 import type { AnthropicProvider as AISDKAnthropicProvider } from '@ai-sdk/anthropic';
 import { logApiRequest } from '../../services/infra/LoggerService';
-import { EventEmitter, EVENT_NAMES } from '../../services/infra/EventEmitter';
 import { hasToolUseTags } from '../../utils/mcpToolParser';
 import { ChunkType, type Chunk } from '../../types/chunk';
-import type { Model, MCPTool } from '../../types';
+import type { Model } from '../../types';
 import { convertMcpToolsToAISDK } from './tools';
 import { supportsExtendedThinking, isClaudeReasoningModel } from './client';
-import { getAppropriateTag, type ReasoningTag, DEFAULT_REASONING_TAGS } from '../../config/reasoningTags';
-
-/**
- * 解析推理标签内容（用于兼容非原生推理模式）
- */
-class ThinkTagParser {
-  private contentBuffer = '';
-  private isInThinkTag = false;
-  private thinkBuffer = '';
-  private reasoningStartTime = 0;
-  private hasReasoningContent = false;
-  private openingTag: string;
-  private closingTag: string;
-
-  constructor(tag?: ReasoningTag) {
-    this.openingTag = tag?.openingTag || '<thinking>';
-    this.closingTag = tag?.closingTag || '</thinking>';
-  }
-
-  processChunk(text: string): { normalText: string; thinkText: string; isThinking: boolean } {
-    this.contentBuffer += text;
-    
-    let normalText = '';
-    let thinkText = '';
-    let processedAny = true;
-
-    while (processedAny && this.contentBuffer.length > 0) {
-      processedAny = false;
-
-      if (!this.isInThinkTag) {
-        const thinkStartIndex = this.contentBuffer.indexOf(this.openingTag);
-        if (thinkStartIndex !== -1) {
-          normalText += this.contentBuffer.substring(0, thinkStartIndex);
-          this.isInThinkTag = true;
-          if (!this.hasReasoningContent) {
-            this.hasReasoningContent = true;
-            this.reasoningStartTime = Date.now();
-          }
-          this.contentBuffer = this.contentBuffer.substring(thinkStartIndex + this.openingTag.length);
-          processedAny = true;
-        } else if (this.contentBuffer.length > this.openingTag.length + 5) {
-          const safeLength = this.contentBuffer.length - (this.openingTag.length + 5);
-          const safeContent = this.contentBuffer.substring(0, safeLength);
-          normalText += safeContent;
-          this.contentBuffer = this.contentBuffer.substring(safeLength);
-          processedAny = true;
-        }
-      } else {
-        const thinkEndIndex = this.contentBuffer.indexOf(this.closingTag);
-        if (thinkEndIndex !== -1) {
-          thinkText += this.contentBuffer.substring(0, thinkEndIndex);
-          this.thinkBuffer += this.contentBuffer.substring(0, thinkEndIndex);
-          this.isInThinkTag = false;
-          this.contentBuffer = this.contentBuffer.substring(thinkEndIndex + this.closingTag.length);
-          processedAny = true;
-        } else if (this.contentBuffer.length > this.closingTag.length + 5) {
-          const safeLength = this.contentBuffer.length - (this.closingTag.length + 5);
-          const safeThinkContent = this.contentBuffer.substring(0, safeLength);
-          thinkText += safeThinkContent;
-          this.thinkBuffer += safeThinkContent;
-          this.contentBuffer = this.contentBuffer.substring(safeLength);
-          processedAny = true;
-        }
-      }
-    }
-
-    return { normalText, thinkText, isThinking: this.isInThinkTag };
-  }
-
-  flush(): { normalText: string; thinkText: string } {
-    let normalText = '';
-    let thinkText = '';
-
-    if (this.contentBuffer.length > 0) {
-      if (this.isInThinkTag) {
-        thinkText = this.contentBuffer;
-        this.thinkBuffer += this.contentBuffer;
-      } else {
-        normalText = this.contentBuffer;
-      }
-      this.contentBuffer = '';
-    }
-
-    return { normalText, thinkText };
-  }
-
-  getFullThinkContent(): string {
-    return this.thinkBuffer;
-  }
-
-  getReasoningTime(): number {
-    return this.hasReasoningContent ? Date.now() - this.reasoningStartTime : 0;
-  }
-}
+import { getAppropriateTag, DEFAULT_REASONING_TAGS } from '../../config/reasoningTags';
+import {
+  ThinkTagParser,
+  accumulateToolCall,
+  emitStreamComplete,
+  emitStreamError,
+  detectAndEmitToolUseTags,
+  PROMPT_MODE_STOP_SEQUENCES,
+  type BaseStreamResult,
+  type BaseStreamParams,
+} from '../../ai/adapters/shared/streamShared';
 
 /**
  * 流式响应结果类型
  */
-export interface StreamResult {
-  content: string;
-  reasoning?: string;
-  reasoningTime?: number;
-  hasToolCalls?: boolean;
-  nativeToolCalls?: any[];
-}
+export interface StreamResult extends BaseStreamResult {}
 
 /**
  * 流式请求参数
  */
-export interface StreamParams {
-  messages?: any[];
-  temperature?: number;
-  max_tokens?: number;
-  top_p?: number;
-  tools?: any[];
-  tool_choice?: any;
-  signal?: AbortSignal;
-  enableTools?: boolean;
-  mcpTools?: MCPTool[];
-  mcpMode?: 'prompt' | 'function';
-  model?: Model;
-  /** 自定义请求体参数 */
-  extraBody?: Record<string, any>;
+export interface StreamParams extends BaseStreamParams {
   /** 是否启用 Extended Thinking */
   enableThinking?: boolean;
   /** 思考预算 Token 数 */
@@ -203,7 +100,7 @@ export async function streamCompletion(
 
     // 🛡️ Prompt 模式防幻觉：添加 stopSequences
     const isPromptMode = !enableTools && mcpTools.length > 0;
-    const stopSequences = isPromptMode ? ['<tool_use_result'] : undefined;
+    const stopSequences = isPromptMode ? PROMPT_MODE_STOP_SEQUENCES : undefined;
 
     // 准备 providerOptions
     let providerOptions: Record<string, any> = {};
@@ -308,26 +205,7 @@ export async function streamCompletion(
 
         case 'tool-call':
           console.log(`[Anthropic SDK Stream] 检测到工具调用: ${part.toolName}`);
-          const toolInput = (part as any).input || (part as any).args || {};
-          toolCalls.push({
-            id: part.toolCallId,
-            toolUseId: part.toolCallId,
-            type: 'function',
-            function: {
-              name: part.toolName,
-              arguments: JSON.stringify(toolInput)
-            }
-          });
-          
-          onChunk?.({
-            type: ChunkType.MCP_TOOL_CREATED,
-            responses: [{
-              id: part.toolCallId,
-              name: part.toolName,
-              arguments: toolInput,
-              status: 'pending'
-            }]
-          });
+          accumulateToolCall(part, toolCalls, onChunk, { includeToolUseId: true });
           break;
 
         case 'finish':
@@ -370,22 +248,10 @@ export async function streamCompletion(
     }
 
     // 发送全局事件
-    EventEmitter.emit(EVENT_NAMES.STREAM_COMPLETE, {
-      provider: 'anthropic-aisdk',
-      model: modelId,
-      content: fullContent,
-      reasoning: fullReasoning,
-      timestamp: Date.now()
-    });
+    emitStreamComplete('anthropic-aisdk', modelId, fullContent, fullReasoning);
 
     // 检查工具使用标签（XML 模式）
-    if (hasToolUseTags(fullContent)) {
-      console.log(`[Anthropic SDK Stream] 检测到 XML 工具使用标签`);
-      EventEmitter.emit(EVENT_NAMES.TOOL_USE_DETECTED, {
-        content: fullContent,
-        model: modelId
-      });
-    }
+    detectAndEmitToolUseTags(fullContent, modelId, 'Anthropic SDK Stream');
 
     const endTime = Date.now();
     console.log(`[Anthropic SDK Stream] 完成，耗时: ${endTime - startTime}ms`);
@@ -400,14 +266,7 @@ export async function streamCompletion(
 
   } catch (error: any) {
     console.error('[Anthropic SDK Stream] 流式响应失败:', error);
-
-    EventEmitter.emit(EVENT_NAMES.STREAM_ERROR, {
-      provider: 'anthropic-aisdk',
-      model: modelId,
-      error: error.message,
-      timestamp: Date.now()
-    });
-
+    emitStreamError('anthropic-aisdk', modelId, error);
     throw error;
   }
 }
@@ -456,7 +315,7 @@ export async function nonStreamCompletion(
 
     // 🛡️ Prompt 模式防幻觉：添加 stopSequences
     const isPromptMode = !enableTools && mcpTools.length > 0;
-    const stopSequences = isPromptMode ? ['<tool_use_result'] : undefined;
+    const stopSequences = isPromptMode ? PROMPT_MODE_STOP_SEQUENCES : undefined;
 
     // 准备 providerOptions
     let providerOptions: Record<string, any> = {};
